Menu: extract the quick (corner/radial) tree into build_quick.cpp
Verbatim move of the quick_out block into build_quick_menu(ctx);
build_menu now assigns its result through ctx.quick_out.
