+++ CMakeLists.txt
@@ -203,6 +203,7 @@ set(SOURCES
     src/hud/hud_renderer.cpp
     src/hud/background_library.cpp
     src/menu/build_menu.cpp
+    src/menu/build_quick.cpp
     src/menu/menu_system.cpp
     src/menu/file_picker.cpp
     src/menu/face_editor.cpp
