+++ src/menu/build_menu.cpp
@@ -9348,213 +9348,8 @@ std::vector<MenuItem> build_menu(MenuBuildContext& ctx)
         "full-setup profiles & backups."));
 
     // ── Quick (corner / radial) menu ─────────────────────────────────────────────
-    // A short, curated set of mid-use actions — separate from the full settings tree
-    // above. Some items are always present; an optional "catalog" of extras can be
-    // pinned by the user (favorites, persisted in config) and is gated by visible_fn.
-    // Submenus here become outer rings in the radial renderer.
-    if (quick_out) {
-        std::vector<MenuItem> q;
-
-        // Night vision (mirror of the Low-Light "Night Vision" toggle).
-        q.push_back(toggle("Night Vision",
-            [&state]{ return state.night_vision.nv_enabled; },
-            [&state](bool v){
-                state.night_vision.nv_enabled  = v;
-                state.night_vision.exposure_ev = v ? 3.0f : 0.0f;
-                state.night_vision.shutter_us  = v ? 40000 : 16667;
-            }));
-
-        // Digital zoom (both eyes together).
-        q.push_back(slider("Zoom", 1.0f, 4.0f, 0.25f, "x",
-            [&state]{ return state.zoom_left.zoom; },
-            [&state](float v){ state.zoom_left.zoom = v; state.zoom_right.zoom = v; }));
-
-        // Manual focus — only when OWLsight cameras are present.
-        {
-            std::vector<MenuItem> focus = {
-                leaf("Autofocus", [cameras]{
-                    if (cameras && cameras->owl_left())  cameras->owl_left()->start_autofocus();
-                    if (cameras && cameras->owl_right()) cameras->owl_right()->start_autofocus();
-                }),
-                slider("Position", 0.f, 1000.f, 25.f, "",
-                    [cameras]{ return (cameras && cameras->owl_left())
-                                        ? (float)cameras->owl_left()->get_focus_position() : 0.f; },
-                    [cameras](float v){
-                        if (cameras && cameras->owl_left())  cameras->owl_left()->set_focus_position((int)v);
-                        if (cameras && cameras->owl_right()) cameras->owl_right()->set_focus_position((int)v);
-                    }),
-            };
-            MenuItem fsub = submenu("Manual Focus", std::move(focus));
-            fsub.visible_fn = [cameras]{ return cameras && cameras->owl_left(); };
-            q.push_back(std::move(fsub));
-        }
-
-        // Quick photo → opens a sub-ring of capture modes; entering it locks focus
-        // (stops AF + manual) so the whole burst shares one focus. All shoot both eyes.
-        {
-            auto shoot = [state_ptr](int extra) {
-                if (!state_ptr) return;
-                std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                state_ptr->capture_request = CaptureRequest::Stereo;
-                state_ptr->capture_burst   = extra;
-            };
-            std::vector<MenuItem> photo = {
-                leaf("Single",   [shoot]{ shoot(0); }),
-                leaf("Burst x3", [shoot]{ shoot(2); }),
-                leaf("Burst",    [shoot]{ shoot(7); }),
-            };
-            MenuItem pm = submenu("Quick Photo", std::move(photo));
-            pm.action = [cameras, state_ptr] {            // on-enter: lock focus
-                if (cameras) {
-                    if (cameras->owl_left())  cameras->owl_left()->stop_autofocus();
-                    if (cameras->owl_right()) cameras->owl_right()->stop_autofocus();
-                }
-                if (state_ptr) {
-                    std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                    state_ptr->focus_left.mode  = CameraFocusState::Mode::MANUAL;
-                    state_ptr->focus_right.mode = CameraFocusState::Mode::MANUAL;
-                }
-            };
-            q.push_back(std::move(pm));
-        }
-        q.push_back(toggle("Record",
-            [&state]{ return state.video_recording; },
-            [&state](bool v){ std::lock_guard<std::mutex> lk(state.mtx);
-                              state.video_request = v ? VideoRequest::Start : VideoRequest::Stop; }));
-
-        // Timers.
-        {
-            auto set_timer = [&state](int secs){
-                state.timer_alarm.timer_active = true;
-                state.timer_alarm.timer_end    = time(nullptr) + secs;
-            };
-            std::vector<MenuItem> timers = {
-                leaf("5 min",  [set_timer]{ set_timer(300);  }),
-                leaf("10 min", [set_timer]{ set_timer(600);  }),
-                leaf("30 min", [set_timer]{ set_timer(1800); }),
-                leaf("60 min", [set_timer]{ set_timer(3600); }),
-                leaf("Cancel", [&state]{ state.timer_alarm.timer_active = false; }),
-            };
-            q.push_back(submenu("Timers", std::move(timers)));
-        }
-
-        // Expand Map — open the Helldivers-style pan/zoom view (closes the wheel).
-        q.push_back(leaf("Expand Map", [state_ptr, menu_sys_pp]{
-            if (state_ptr) {
-                std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                state_ptr->map_overlay.expanded   = true;
-                state_ptr->map_overlay.view_zoom  = 1.f;
-                state_ptr->map_overlay.view_pan_x = 0.f;
-                state_ptr->map_overlay.view_pan_y = 0.f;
-            }
-            if (menu_sys_pp && *menu_sys_pp) (*menu_sys_pp)->close();
-        }));
-
-        // ── Favorites catalog (optional, user-pinned) ────────────────────────────
-        struct QuickFav { std::string key; MenuItem item; };
-        std::vector<QuickFav> catalog;
-        catalog.push_back({ "edge", toggle("Edge Highlight",
-            [&state]{ return state.pp_cfg.edge_enabled; },
-            [&state](bool v){ state.pp_cfg.edge_enabled = v; }) });
-        catalog.push_back({ "desat", toggle("Desaturate BG",
-            [&state]{ return state.pp_cfg.desat_enabled; },
-            [&state](bool v){ state.pp_cfg.desat_enabled = v; }) });
-        catalog.push_back({ "motion", toggle("Motion Highlight",
-            [&state]{ return state.pp_cfg.motion_enabled; },
-            [&state](bool v){ state.pp_cfg.motion_enabled = v; }) });
-        catalog.push_back({ "map", toggle("Map Overlay",
-            [&state]{ return state.map_overlay.enabled; },
-            [&state](bool v){ state.map_overlay.enabled = v; }) });
-        catalog.push_back({ "theater", toggle("Theater Mode",
-            [&state]{ return state.theater_mode; },
-            [&state](bool v){ state.theater_mode = v; }) });
-        catalog.push_back({ "swap", toggle("Swap Cameras",
-            [&state]{ return state.cameras_swapped; },
-            [&state](bool v){ state.cameras_swapped = v; }) });
-        catalog.push_back({ "fps", toggle("FPS Overlay",
-            [fps_overlay_active]{ return fps_overlay_active && *fps_overlay_active; },
-            [fps_overlay_active](bool v){ if (fps_overlay_active) *fps_overlay_active = v; }) });
-        catalog.push_back({ "syspanel", toggle("System Panel",
-            [sys_panel_active]{ return sys_panel_active && *sys_panel_active; },
-            [sys_panel_active](bool v){ if (sys_panel_active) *sys_panel_active = v; }) });
-        // Action item (not a toggle): rings the paired phone via KDE Connect.
-        catalog.push_back({ "ring_phone", leaf("Ring Phone", [kdc_p, &state]{
-            const bool ok = kdc_p && kdc_p->ring_phone();
-            std::lock_guard<std::mutex> lk(state.mtx);
-            Notification n; n.type = NotifType::App; n.icon = "message";
-            n.title = ok ? "Ringing phone\xE2\x80\xA6" : "Phone not connected";
-            n.body  = ok ? "KDE Connect \xC2\xB7 findmyphone"
-                         : "Pair a device in the KDE Connect app first";
-            n.auto_dismiss_s = 4.f;
-            state.notifs.push(std::move(n));
-        }) });
-        // Diagnostics — quick recovery actions (CSI reinit + face-renderer restart).
-        {
-            std::vector<MenuItem> diag;
-            diag.push_back(with_desc(leaf("Reinit CSI Cameras", [cameras, &state]{
-                const bool ok = cameras->reinit_owls();
-                const bool lok = cameras->owl_left_ok(), rok = cameras->owl_right_ok();
-                std::lock_guard<std::mutex> lk(state.mtx);
-                Notification n; n.type = NotifType::App;
-                n.title = ok ? "CSI cameras reinitialized" : "CSI reinit: no camera";
-                char b[64]; snprintf(b, sizeof(b), "Left %s  \xC2\xB7  Right %s",
-                                     lok ? "OK" : "\xE2\x80\x94", rok ? "OK" : "\xE2\x80\x94");
-                n.body = b; n.auto_dismiss_s = 5.f; state.notifs.push(std::move(n));
-            }), "Re-enumerate + restart the CSI cameras to recover a dark eye."));
-            if (pf_restart_renderer) {
-                MenuItem rf = with_desc(leaf("Restart Face Renderer", [pf_restart_renderer, state_ptr]{
-                    pf_restart_renderer();
-                    if (!state_ptr) return;
-                    std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                    Notification n; n.type = NotifType::App; n.title = "Face renderer restarted";
-                    n.body = "Relaunched the HUB75 panel driver"; n.auto_dismiss_s = 4.f;
-                    state_ptr->notifs.push(std::move(n));
-                }), "Kill + relaunch the HUB75 panel pusher to recover the face feed.");
-                rf.visible_fn = [pf_backend_p]{ return !pf_backend_p || *pf_backend_p == "hub75"; };
-                diag.push_back(std::move(rf));
-            }
-            catalog.push_back({ "diagnostics", submenu("Diagnostics", std::move(diag)) });
-        }
-
-        // Pinned catalog items appear in the quick menu, gated on the favorites set.
-        for (auto& f : catalog) {
-            MenuItem it  = f.item;
-            std::string key = f.key;
-            it.visible_fn = [state_ptr, key]{
-                if (!state_ptr) return false;
-                std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                return state_ptr->quick_favorites.count(key) > 0;
-            };
-            q.push_back(std::move(it));
-        }
-
-        // "Customize" — toggles to pin/unpin each catalog item.
-        std::vector<MenuItem> customize;
-        for (auto& f : catalog) {
-            std::string key   = f.key;
-            std::string label = f.item.label;
-            customize.push_back(toggle(label,
-                [state_ptr, key]{
-                    if (!state_ptr) return false;
-                    std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                    return state_ptr->quick_favorites.count(key) > 0;
-                },
-                [state_ptr, key](bool v){
-                    if (!state_ptr) return;
-                    std::lock_guard<std::mutex> lk(state_ptr->mtx);
-                    if (v) state_ptr->quick_favorites.insert(key);
-                    else   state_ptr->quick_favorites.erase(key);
-                }));
-        }
-        q.push_back(submenu("Customize", std::move(customize)));
-
-        // Jump to the full-screen settings.
-        q.push_back(leaf("Full Settings", [menu_sys_pp]{
-            if (menu_sys_pp && *menu_sys_pp) (*menu_sys_pp)->open_deep();
-        }));
-
-        *quick_out = std::move(q);
-    }
+    // Built in src/menu/build_quick.cpp; assigned through quick_out when set.
+    if (quick_out) *quick_out = build_quick_menu(ctx);
 
     return {
         with_desc(submenu("Vision",       std::move(cameras_menu)),
